+++ base_aux/valid/valid_1_base.py
@@ -3,6 +3,7 @@ from base_aux.base_argskwargs import *
 from base_aux.funcs import *
 from base_aux.lambdas.lambdas import *
 from base_aux.funcs.static import TYPE__VALID_VALIDATOR
+from base_aux.cmp.eq import Eq
 
 
 # =====================================================================================================================
@@ -223,7 +224,7 @@ class Valid:
                     self.validate_last = Lambda(self.VALIDATE_LINK).get_result_or_exx(self.value_last, *self.ARGS__VALIDATE, **self.KWARGS__VALIDATE)
 
                 else:
-                    self.validate_last = ValidAux.eq_doublesided_or_exx(self.value_last, self.VALIDATE_LINK)
+                    self.validate_last = Eq.eq_doublesided_or_exx(self.value_last, self.VALIDATE_LINK)
 
                 # FINISH retry
                 if not self.VALIDATE_RETRY or retry_count == self.VALIDATE_RETRY or self.validate_last_bool: