+++ base_aux/valid/value_2_unit.py
@@ -7,7 +7,7 @@ from base_aux.base_argskwargs.novalue import NoValue
 # from base_aux.attrs.annot_4_cls_keys_as_values import AnnotClsKeysAsValues
 from base_aux.classes.number import NumberArithmTranslateToAttr
 
-from base_aux.valid.valid_0_aux import ValidAux
+from base_aux.cmp.eq import Eq
 
 
 # =====================================================================================================================
@@ -257,7 +257,7 @@ class ValueUnit(NumberArithmTranslateToAttr):
             0=self==other
             -1=self<other
         """
-        if ValidAux.eq_doublesided__bool(self.VALUE, other):
+        if Eq.eq_doublesided__bool(self.VALUE, other):
             return 0
         if not isinstance(other, self.__class__):
             other = self.__class__(other)