+++ tests/valid/test__valid_0_aux.py
@@ -2,7 +2,7 @@ from typing import *
 import pytest
 
 from base_aux.pytester import *
-from base_aux.valid.valid_0_aux import ValidAux
+from base_aux.cmp.eq import Eq
 from base_aux.base_objects import *
 
 
@@ -24,13 +24,13 @@ from base_aux.base_objects import *
     ]
 )
 def test__compare_doublesided(args, _EXPECTED):
-    func_link = ValidAux.eq_doublesided_or_exx
+    func_link = Eq.eq_doublesided_or_exx
     pytest_func_tester__no_kwargs(func_link, args, _EXPECTED[0])
 
-    func_link = ValidAux.eq_doublesided__bool
+    func_link = Eq.eq_doublesided__bool
     pytest_func_tester__no_kwargs(func_link, args, _EXPECTED[1])
 
-    func_link = ValidAux.eq_doublesided__reverse
+    func_link = Eq.eq_doublesided__reverse
     pytest_func_tester__no_kwargs(func_link, args, _EXPECTED[2])